+++ src/ATK/structures/Spectrum.py
@@ -1,29 +1,56 @@
 from dataclasses import dataclass, field
+from typing import Self
 
 import astropy.units as u
 import numpy
 from astropy.coordinates import SkyCoord
 from astropy.units import Quantity
+from pandas import DataFrame
 
+from ..utilities.docstrings import get_docstring
 from .methods.spectrum.fitting import do_fitting
 from .methods.spectrum.radial_velocities import get_rvs
 from .structures_core import Container, manage_inplace
+from .Target import Target
 
 
 @dataclass(repr=False)
 class Spectrum(Container):
-    # --- metadata ---
+    """
+    Container for storing spectral data. This object stores both data and relevant metadata.
+
+    .. rubric:: Valid Configurations
+
+    A :class:`~ATK.Models.Spectrum` must be initialised with one of the following mutually exclusive parameters:
+
+    - ``wavelength`` or ``velocity``
+
+    Providing both or neither raises a ``ValueError``.
+    """
+
+    #: DOC_OVERRIDE
     survey: str | None = None
+    #: DOC_OVERRIDE
     correction: str | None = None
+    #: DOC_OVERRIDE
     search_pos: SkyCoord | None = None
+    #: DOC_OVERRIDE
     separation: Quantity | None = None
+    #: Exposure of spectrum.
     exposure: Quantity | None = None
+    #: Reference wavelength.
+    #:
+    #: ``None`` unless spectrum has been converted to a velocity spectrum via :meth:`~ATK.Models.Spectrum.vspec`.
     wav_ref: Quantity | None = None
 
-    # --- data ---
-    wavelength: numpy.ndarray | Quantity | None = None
-    velocity: numpy.ndarray | Quantity | None = None
-    flux: numpy.ndarray | Quantity | None = None
+    #: Wavelength values.
+    #: Mutually exclusive with ``velocity``.
+    wavelength: Quantity | None = None
+    #: Velocity values.
+    #: Mutually exclusive with ``wavelength``.
+    velocity: Quantity | None = None
+    #: Flux values.
+    flux: Quantity | None = None
 
     _required = ["survey"]
 
@@ -66,7 +93,7 @@ class Spectrum(Container):
                 setattr(self, attr, val * unit)
 
     @property
-    def x_type(self):
+    def _x_type(self):
         if self.wavelength is not None:
             return "wavelength"
         elif self.velocity is not None:
@@ -75,11 +102,11 @@ class Spectrum(Container):
             raise ValueError("Spectrum container must hold one of 'wavelength' and 'velocity'.")
 
     @property
-    def x_arr(self):
-        return getattr(self, self.x_type)
+    def _x_arr(self):
+        return getattr(self, self._x_type)
 
-    def set_x(self, val: numpy.ndarray):
-        setattr(self, self.x_type, val)
+    def _set_x(self, val: numpy.ndarray):
+        setattr(self, self._x_type, val)
 
     def crop(self, min: float | None = None, max: float | None = None, inplace=True):
         from .methods.cropping import crop_nd
@@ -88,12 +115,14 @@ class Spectrum(Container):
 
         ys = [struct.flux]
 
-        x, ys = crop_nd(x=struct.x_arr, ys=ys, lower_lim=min, upper_lim=max)
+        x, ys = crop_nd(x=struct._x_arr, ys=ys, lower_lim=min, upper_lim=max)
 
-        struct.set_x(x)
+        struct._set_x(x)
         struct.flux = ys[0]
 
         return struct
+
+    crop.__doc__ = get_docstring("bin", x="``wavelength`` or ``velocity``", name="Spectrum")
 
     def bin(self, bins: int | None = None, size: Quantity | float | None = None, inplace=True):
         from .methods.binning import bin_nd
@@ -102,14 +131,31 @@ class Spectrum(Container):
 
         ys = [struct.flux]
 
-        x, ys, _ = bin_nd(x=struct.x_arr, ys=ys, errs=[], bins=bins, size=size)
+        x, ys, _ = bin_nd(x=struct._x_arr, ys=ys, errs=[], bins=bins, size=size)
 
-        struct.set_x(x)
+        struct._set_x(x)
         struct.flux = ys[0]
 
         return struct
 
+    bin.__doc__ = get_docstring("bin", x="``wavelength`` or ``velocity``", name="Spectrum")
+
     def vspec(self, wav_ref: float | Quantity, inplace: bool = True):
+        """
+        Convert wavelength values to velocity relative to a reference wavelength.
+        This replaces the stored ``wavelength`` axis with ``velocity``.
+
+        Parameters
+        ----------
+        wav_ref : float | Quantity
+            Reference (rest) wavelength used to compute velocities.
+
+            If a :class:`~astropy.units.Unit` is not provided, ``wav_ref`` is assumed to be in the same unit as ``wavelength``.
+
+        inplace : bool, optional
+            If ``True``, modify the current :class:`~ATK.Models.Spectrum` in place - leaving the original unchanged. If ``False``, operate on and return a copy.
+        """
+
         from .methods.spectrum.radial_velocities import get_velocities
 
         struct = manage_inplace(self, inplace)
@@ -122,3 +168,15 @@ class Spectrum(Container):
         struct.wavelength = None
 
         return struct
+
+    @classmethod
+    def from_dataframe(cls, target: Target | int | SkyCoord, data: DataFrame, **kwargs) -> Self:
+        return super().from_dataframe(target, data, **kwargs)
+
+    from_dataframe.__func__.__doc__ = get_docstring("from_dataframe", obj="Spectrum", args=", ".join(f"``{p}``" for p in _required))
+
+    @classmethod
+    def from_table(cls, target: Target | int | SkyCoord, data: DataFrame, **kwargs) -> Self:
+        return super().from_table(target, data, **kwargs)
+
+    from_table.__func__.__doc__ = get_docstring("from_table", obj="Spectrum", args=", ".join(f"``{p}``" for p in _required))